+++ app/src/main/java/com/vietinterview/getbee/api/response/detailjob/DetailJobResponse.java
@@ -43,7 +43,7 @@ public class DetailJobResponse implements Parcelable {
     private Integer currency;
     @SerializedName("fee")
     @Expose
-    private Integer fee;
+    private Long fee;
     @SerializedName("jobDescription")
     @Expose
     private String jobDescription;
@@ -159,7 +159,7 @@ public class DetailJobResponse implements Parcelable {
         if (in.readByte() == 0) {
             fee = null;
         } else {
-            fee = in.readInt();
+            fee = in.readLong();
         }
         jobDescription = in.readString();
         submitDate = in.readString();
@@ -349,11 +349,11 @@ public class DetailJobResponse implements Parcelable {
         this.currency = currency;
     }
 
-    public Integer getFee() {
+    public Long getFee() {
         return fee;
     }
 
-    public void setFee(Integer fee) {
+    public void setFee(Long fee) {
         this.fee = fee;
     }
 
@@ -578,7 +578,7 @@ public class DetailJobResponse implements Parcelable {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeInt(fee);
+            parcel.writeLong(fee);
         }
         parcel.writeString(jobDescription);
         parcel.writeString(submitDate);
