+++ app/src/main/java/com/vietinterview/getbee/api/response/jobs/JobList.java
@@ -41,7 +41,7 @@ public class JobList implements Parcelable {
     private Integer currency;
     @SerializedName("fee")
     @Expose
-    private Integer fee;
+    private Long fee;
     @SerializedName("jobDescription")
     @Expose
     private String jobDescription;
@@ -143,7 +143,7 @@ public class JobList implements Parcelable {
         if (in.readByte() == 0) {
             fee = null;
         } else {
-            fee = in.readInt();
+            fee = in.readLong();
         }
         jobDescription = in.readString();
         submitDate = in.readString();
@@ -292,11 +292,11 @@ public class JobList implements Parcelable {
         this.currency = currency;
     }
 
-    public Integer getFee() {
+    public Long getFee() {
         return fee;
     }
 
-    public void setFee(Integer fee) {
+    public void setFee(Long fee) {
         this.fee = fee;
     }
 
@@ -497,7 +497,7 @@ public class JobList implements Parcelable {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeInt(fee);
+            parcel.writeLong(fee);
         }
         parcel.writeString(jobDescription);
         parcel.writeString(submitDate);
